requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
openai
PyPDF2
python-docx
SpeechRecognition
pydub
ffmpeg-python